+++ src/__tests__/standalone-test.js
@@ -13,7 +13,11 @@ const fixturesPath = path.join(__dirname, "fixtures")
 test("standalone with input file(s)", t => {
   let planned = 0
 
-  standalone({ files: `${fixturesPath}/empty-block.css`, config: configBlockNoEmpty })
+  standalone({
+    files: `${fixturesPath}/empty-block.css`,
+    // Path to config file
+    configFile: path.join(__dirname, "fixtures/config-block-no-empty.json"),
+  })
     .then(({ output, results }) => {
       t.ok(output.indexOf("block-no-empty") !== -1)
       t.equal(results.length, 1)